refactor: move deck filtering rules to entity
