+++ src/entities/card/@x/deck.ts
@@ -1,1 +1,2 @@
 export { deleteLocalCardsByDeckId } from "../model/store";
+export type { Card } from "../model/types";
