+++ src/entities/deck/index.ts
@@ -1,7 +1,14 @@
 export { fetchDecks, subscribeDecks } from "./api/firestore";
 export { createDeck, deleteDeck, editDeck } from "./api/mutations";
 export { generateDeckId } from "./model/id";
-export { CATEGORY, getCategory, isHighlightLanguage, mustFindDeckById } from "./model/rules";
+export {
+  CATEGORY,
+  createSelectableStudyCard,
+  filterCardsForDeck,
+  getCategory,
+  isHighlightLanguage,
+  mustFindDeckById,
+} from "./model/rules";
 export { useDeck, useDecks } from "./model/hooks";
 export { clearRemoteDecks } from "./model/store";
 export type { Deck, DeckCreateInput, DeckEdit, DeckId, RemoteDeck } from "./model/types";
