+++ src/entities/study-progress/@x/deck.ts
@@ -1,0 +1,2 @@
+export { createStudyProgressFromCard, isStudyProgressEligible } from "../model/rules";
+export type { StudyProgress } from "../model/types";
