+++ src/entities/study-progress/index.ts
@@ -3,7 +3,6 @@ export {
   compareStudyProgress,
   createStudyProgressFromCard,
   getNextStudyAvailabilityAt,
-  isStudyProgressEligible,
   recordStudyProgress,
 } from "./model/rules";
 export type { StudyProgress, StudyProgressEdit, StudyRating } from "./model/types";
