--- src/features/deck-filter/model/cardSelection.ts
@@ -1,47 +1,0 @@
-import type { Card } from "@/entities/card";
-import type { Deck } from "@/entities/deck";
-import type { StudyPreferences } from "@/entities/preferences";
-import { createStudyProgressFromCard, isStudyProgressEligible, type StudyProgress } from "@/entities/study-progress";
-
-interface SelectableStudyCard<TCard extends Card = Card> {
-  card: TCard;
-  progress: StudyProgress;
-}
-
-export const createSelectableStudyCard = <TCard extends Card>(card: TCard): SelectableStudyCard<TCard> => ({
-  card,
-  progress: createStudyProgressFromCard({
-    id: card.id,
-    score: card.score,
-    numberOfSeen: card.numberOfSeen,
-    ...(card.lastSeenAt === undefined ? {} : { lastSeenAt: card.lastSeenAt }),
-    ...(card.nextSeeingAt === undefined ? {} : { nextSeeingAt: card.nextSeeingAt }),
-    ...(card.interval === undefined ? {} : { interval: card.interval }),
-  }),
-});
-
-const isCardMatchingTags = (card: Card, deck: Pick<Deck, "selectedTags" | "tagAndFilter">) => {
-  const tags = deck.selectedTags;
-  if (tags.length === 0) return true;
-  if (deck.tagAndFilter) return tags.every((tag) => card.tags.includes(tag));
-  return tags.some((tag) => card.tags.includes(tag));
-};
-
-export const filterCardsForDeck = <TCard extends Card>(
-  cards: SelectableStudyCard<TCard>[],
-  deck: Pick<Deck, "selectedTags" | "tagAndFilter" | "scoreMax" | "scoreMin">,
-  study: Pick<StudyPreferences, "useCardInterval">,
-  now: number
-): SelectableStudyCard<TCard>[] =>
-  cards.filter(({ card, progress }) => {
-    if (!isCardMatchingTags(card, deck)) return false;
-    return isStudyProgressEligible(
-      progress,
-      {
-        maximumScore: deck.scoreMax,
-        minimumScore: deck.scoreMin,
-        respectNextSeeingAt: study.useCardInterval,
-      },
-      now
-    );
-  });
